# docs/swagger.yaml
basePath: /
definitions:
...
  models.View:
    properties:
      createdAt:
        type: string
      filterConfig:
        additionalProperties: {}
        type: object
      id:
        type: integer
      name:
        type: string
      searchKeywordURLID:
        type: integer
      updatedAt:
        type: string
    type: object
...
  views.CreateViewData:
    properties:
      filter_config:
        additionalProperties: {}
        type: object
      name:
        type: string
      search_keyword_url_id:
        type: integer
    type: object
  views.CreateViewRequest:
    properties:
      data:
        $ref: '#/definitions/views.CreateViewData'
    type: object
  views.DeleteViewData:
    properties:
      id:
        type: integer
    type: object
  views.DeleteViewResponse:
    properties:
      data:
        $ref: '#/definitions/views.DeleteViewData'
    type: object
  views.UpdateViewData:
    properties:
      filter_config:
        additionalProperties: {}
        type: object
      name:
        type: string
    type: object
  views.UpdateViewRequest:
    properties:
      data:
        $ref: '#/definitions/views.UpdateViewData'
      id:
        type: integer
    type: object
  views.ViewPageCountData:
    properties:
      page_count:
        type: integer
    type: object
  views.ViewPageCountResponse:
    properties:
      data:
        $ref: '#/definitions/views.ViewPageCountData'
    type: object
  views.ViewResponse:
    properties:
      data:
        $ref: '#/definitions/models.View'
    type: object
  views.ViewsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.View'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Access token for authentication
  termsOfService: http://swagger.io/terms/
  title: SiteCrawler API
  version: "1.0"
paths:
...
  /api/views:
    get:
      description: Lists views for a given `search_keyword_url_id`
      parameters:
      - description: Search keyword URL ID
        in: query
        name: search_keyword_url_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.ViewsResponse'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List views
      tags:
      - Views
    post:
      consumes:
      - application/json
      description: Creates a new view
      parameters:
      - description: View payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/views.CreateViewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.ViewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create view
      tags:
      - Views
  /api/views/{id}:
    delete:
      description: Deletes a view by ID
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.DeleteViewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete view
      tags:
      - Views
    get:
      description: Fetches a view by ID
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.ViewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get view
      tags:
      - Views
    put:
      consumes:
      - application/json
      description: Updates a view by ID
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: integer
      - description: View update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/views.UpdateViewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.ViewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update view
      tags:
      - Views
  /api/views/{id}/page_count:
    get:
      description: Returns a count of pages for the given view and optional crawling
        session filter
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: integer
      - description: View ID (must match path)
        in: query
        name: view_id
        required: true
        type: integer
      - description: Crawling session ID
        in: query
        name: crawling_session_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.ViewPageCountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: View page count
      tags:
      - Views
...
securityDefinitions:
  ApiKeyAuth:
    description: Access token for authentication
    in: header
    name: X-Access-Token
    type: apiKey
swagger: "2.0"